mask_rcnn/requirements.txt
lxml
matplotlib
numpy
tqdm
pycocotools
Pillow
torch==1.10
torchvision==0.11.1
